refactor assets
